requirements.txt
streamlit
openai>=1.0.0
pandas
requests
python-dotenv
numpy
python-dateutil
tqdm